Split CLI filter parsing into dedicated module
